:recycle:  Extract test for GF events
