+++ src/infra/sequelize/queries/frise/getProjectEvents.integration.ts
@@ -10,7 +10,6 @@ import makeFakeProject from '../../../../__tests__/fixtures/project'
 describe('frise.getProjectEvents', () => {
   const eventTimestamp = new Date('2022-01-04').getTime()
   const notifiedOnTimestamp = new Date('2022-01-05').getTime()
-  const GFDueDateTimestamp = new Date('2022-03-05').getTime()
 
   const { Project } = models
   const projectId = new UniqueEntityID().toString()
@@ -72,7 +71,7 @@ describe('frise.getProjectEvents', () => {
   for (const role of USER_ROLES.filter((role) => role !== 'ademe')) {
     describe(`when the user is ${role}`, () => {
       const fakeUser = { role } as User
-      it('should return the ProjectNotified and ProjectGFDueDateSet events', async () => {
+      it('should return the ProjectNotified events', async () => {
         await ProjectEvent.create({
           id: new UniqueEntityID().toString(),
           projectId,
@@ -80,14 +79,6 @@ describe('frise.getProjectEvents', () => {
           valueDate: notifiedOnTimestamp,
           eventPublishedAt: eventTimestamp,
         })
-        await ProjectEvent.create({
-          id: new UniqueEntityID().toString(),
-          projectId,
-          type: 'ProjectGFDueDateSet',
-          valueDate: GFDueDateTimestamp,
-          eventPublishedAt: eventTimestamp,
-        })
-
         const res = await getProjectEvents({ projectId, user: fakeUser })
 
         expect(res._unsafeUnwrap()).toMatchObject({
@@ -95,11 +86,6 @@ describe('frise.getProjectEvents', () => {
             {
               type: 'ProjectNotified',
               date: notifiedOnTimestamp,
-              variant: role,
-            },
-            {
-              type: 'ProjectGFDueDateSet',
-              date: GFDueDateTimestamp,
               variant: role,
             },
           ],
@@ -138,19 +124,12 @@ describe('frise.getProjectEvents', () => {
 
   describe(`when the user is ademe`, () => {
     const fakeUser = { role: 'ademe' } as User
-    it('should not return the ProjectNotified and ProjectGFDueDateSet events', async () => {
+    it('should not return the ProjectNotified events', async () => {
       await ProjectEvent.create({
         id: new UniqueEntityID().toString(),
         projectId,
         type: 'ProjectNotified',
         valueDate: eventTimestamp,
-        eventPublishedAt: eventTimestamp,
-      })
-      await ProjectEvent.create({
-        id: new UniqueEntityID().toString(),
-        projectId,
-        type: 'ProjectGFDueDateSet',
-        valueDate: GFDueDateTimestamp,
         eventPublishedAt: eventTimestamp,
       })
 
@@ -299,137 +278,6 @@ describe('frise.getProjectEvents', () => {
         expect(res._unsafeUnwrap()).toMatchObject({
           events: [],
         })
-      })
-    })
-  }
-
-  for (const role of USER_ROLES.filter(
-    (role) => role === 'porteur-projet' || role === 'admin' || role === 'dgec' || role === 'dreal'
-  )) {
-    const fakeUser = { role } as User
-    describe(`when user is ${role}`, () => {
-      it('should return ProjectGFSubmitted, ProjectGFRemoved, ProjectGFValidated, and ProjectGFInvalidated events', async () => {
-        const fileId = new UniqueEntityID().toString()
-        const gfDate = new Date('2021-12-26').getTime()
-        const removedAt = new Date('2021-12-30').getTime()
-        const validatedAt = new Date('2022-01-14').getTime()
-        const invalidatedAt = new Date('2022-01-15').getTime()
-
-        await ProjectEvent.create({
-          id: new UniqueEntityID().toString(),
-          projectId,
-          type: 'ProjectGFSubmitted',
-          valueDate: gfDate,
-          eventPublishedAt: new Date('2021-12-27').getTime(),
-          payload: {
-            fileId: fileId,
-            filename: 'my-file',
-          },
-        })
-
-        await ProjectEvent.create({
-          id: new UniqueEntityID().toString(),
-          projectId,
-          type: 'ProjectGFRemoved',
-          valueDate: removedAt,
-          eventPublishedAt: removedAt,
-        })
-
-        await ProjectEvent.create({
-          type: 'ProjectGFValidated',
-          projectId,
-          valueDate: validatedAt,
-          eventPublishedAt: validatedAt,
-          id: new UniqueEntityID().toString(),
-        })
-
-        await ProjectEvent.create({
-          type: 'ProjectGFInvalidated',
-          projectId,
-          valueDate: invalidatedAt,
-          eventPublishedAt: invalidatedAt,
-          id: new UniqueEntityID().toString(),
-        })
-
-        const res = await getProjectEvents({ projectId, user: fakeUser })
-
-        expect(res._unsafeUnwrap().events).toHaveLength(4)
-        expect(res._unsafeUnwrap()).toMatchObject({
-          events: [
-            {
-              type: 'ProjectGFSubmitted',
-              date: gfDate,
-              variant: role,
-              fileId: fileId,
-              filename: 'my-file',
-            },
-            {
-              type: 'ProjectGFRemoved',
-              date: removedAt,
-              variant: role,
-            },
-            {
-              type: 'ProjectGFValidated',
-              date: validatedAt,
-              variant: role,
-            },
-            {
-              type: 'ProjectGFInvalidated',
-              date: invalidatedAt,
-              variant: role,
-            },
-          ],
-        })
-      })
-    })
-  }
-
-  for (const role of USER_ROLES.filter(
-    (role) => role !== 'porteur-projet' && role !== 'admin' && role !== 'dgec' && role !== 'dreal'
-  )) {
-    const fakeUser = { role } as User
-    describe(`when user is ${role}`, () => {
-      it('should NOT return ProjectGFSubmitted, ProjectGFRemoved and, ProjectGFInvalidated and ProjectGFValidated for GF events', async () => {
-        await ProjectEvent.create({
-          id: new UniqueEntityID().toString(),
-          projectId,
-          type: 'ProjectGFSubmitted',
-          valueDate: new Date('2021-12-26').getTime(),
-          eventPublishedAt: new Date('2021-12-27').getTime(),
-          payload: {
-            fileId: new UniqueEntityID().toString(),
-            filename: 'my-file',
-          },
-        })
-
-        await ProjectEvent.create({
-          id: new UniqueEntityID().toString(),
-          projectId,
-          type: 'ProjectGFRemoved',
-          valueDate: new Date('2021-12-30').getTime(),
-          eventPublishedAt: new Date('2021-12-30').getTime(),
-        })
-
-        await ProjectEvent.create({
-          type: 'ProjectGFValidated',
-          projectId,
-          valueDate: new Date('2021-12-30').getTime(),
-          eventPublishedAt: new Date('2021-12-30').getTime(),
-          id: new UniqueEntityID().toString(),
-        })
-
-        await ProjectEvent.create({
-          type: 'ProjectGFInvalidated',
-          projectId,
-          valueDate: new Date('2021-12-30').getTime(),
-          eventPublishedAt: new Date('2021-12-30').getTime(),
-          id: new UniqueEntityID().toString(),
-        })
-
-        const res = await getProjectEvents({ projectId, user: fakeUser })
-
-        expect(res._unsafeUnwrap().events).toHaveLength(0)
-        expect(res._unsafeUnwrap()).toMatchObject({ events: [] })
       })
     })
   }
